ATS/requirements.txt
djangorestframework
openpyxl
pandas


pip install djangorestframework openpyxl pandas